[FEAT] Move Crop Circle to Craftables
